Use the endpoint read stream interface, introduced in gousb 2.0. (#41)
* Use the endpoint read stream interface, introduced in gousb 2.0.

* With the streaming interface, we can keep multiple transfers queued
up on the host and waiting for the data. With a few transfers waiting
for data all the time, the size of the buffer no longer matters that
much and we can use smaller buffers, and don't need to pay particular
attention to it's size. Use a buffer large enough for 10 ISO packets
(or 60 bulk packets).

* Merge branch 'master' into gousb_stream

* Add comments about picked buffer/queue sizes.